--- test/AdoNetCore.AseClient.Tests/Integration/InternalConnectionTests.cs
@@ -1,25 +1,0 @@
-﻿using AdoNetCore.AseClient.Internal;
-using NUnit.Framework;
-
-namespace AdoNetCore.AseClient.Tests.Integration
-{
-    [TestFixture]
-    [Category("basic")]
-    public class InternalConnectionTests
-    {
-        public InternalConnectionTests()
-        {
-            Logger.Enable();
-        }
-
-        [Test]
-        public void Ping_ShouldWork()
-        {
-            using (var connection = new AseConnection(ConnectionStrings.Default))
-            {
-                connection.Open();
-                Assert.IsTrue(connection.InternalConnection.Ping());
-            }
-        }
-    }
-}
